Community-platform-main/backend_code/requirements.txt
Django>=3.2
djangorestframework>=3.12
mysqlclient>=2.0
django-cors-headers>=3.7
djangorestframework-simplejwt>=5.0
python-dotenv>=0.19.1
Pillow>=8.0.0  # For image handling
python-magic>=0.4.22  # For file type checking
django-filter>=21.1  # For filtering
requests>=2.26.0  # For external API requests
drf-yasg>=1.20.0  # For API documentation
pyyaml>=6.0  # Required by drf-yasg